// client/src/app/@types/auth.schema.ts
import { Role } from '../constants'
import z from 'zod'

export const LoginBody = z
  .object({
    email: z.string().min(1, { message: 'required' }).email({
      message: 'invalidEmail'
    }),
    password: z.string().min(6, 'minmaxPassword').max(100, 'minmaxPassword')
  })
  .strict()

export type LoginBodyType = z.TypeOf<typeof LoginBody>

export const LoginRes = z.object({
  data: z.object({
    accessToken: z.string(),
    refreshToken: z.string(),
    account: z.object({
      id: z.number(),
      name: z.string(),
      email: z.string(),
      role: z.enum([Role.Owner, Role.Employee])
    })
  }),
  message: z.string()
})

export type LoginResType = z.TypeOf<typeof LoginRes>


export const RegisterBody = z
  .object({
    email: z.string().min(1, { message: 'required' }).email({
      message: 'invalidEmail'
    }),
    password: z.string().min(6, 'minmaxPassword').max(100, 'minmaxPassword'),
    displayName: z.string().min(6, 'minmaxName').max(50, 'minmaxName'),
  })
  .strict()

export type RegisterBodyType = z.TypeOf<typeof RegisterBody>
export const RegisterRes = z.object({
  accessToken: z.string(),
  refreshToken: z.string(),
  refreshExpiresAt: z.string(),
  accessExpiresAt: z.string()
})
export type RegisterResType = z.TypeOf<typeof RegisterRes>

export const RefreshTokenBody = z
  .object({
    refreshToken: z.string()
  })
  .strict()

export type RefreshTokenBodyType = z.TypeOf<typeof RefreshTokenBody>

export const RefreshTokenRes = z.object({
  data: z.object({
    accessToken: z.string(),
    refreshToken: z.string()
  }),
  message: z.string()
})

export type RefreshTokenResType = z.TypeOf<typeof RefreshTokenRes>

export const LogoutBody = z
  .object({
    refreshToken: z.string()
  })
  .strict()

export type LogoutBodyType = z.TypeOf<typeof LogoutBody>